+++ core/src/main/scala/at/logic/gapt/formats/tip/problem.scala
@@ -2,7 +2,7 @@ package at.logic.gapt.formats.tip
 
 import at.logic.gapt.expr._
 import at.logic.gapt.expr.hol.{ existentialClosure, universalClosure }
-import at.logic.gapt.proofs.{ Context, Sequent }
+import at.logic.gapt.proofs.{ Context, ImmutableContext, Sequent }
 
 case class TipConstructor( constr: Const, projectors: Seq[Const] ) {
   val FunctionType( datatype, fieldTypes ) = constr.ty
@@ -23,7 +23,7 @@ case class TipDatatype( t: TBase, constructors: Seq[TipConstructor] ) {
 case class TipFun( fun: Const, definitions: Seq[Formula] )
 
 case class TipProblem(
-    ctx:   Context,
+    ctx:   ImmutableContext,
     sorts: Seq[TBase], datatypes: Seq[TipDatatype],
     uninterpretedConsts: Seq[Const], functions: Seq[TipFun],
     assumptions: Seq[Formula], goal: Formula ) {
@@ -48,7 +48,7 @@ case class TipProblem(
       Sequent()
       :+ goal )
 
-  def context: Context = ctx
+  def context: ImmutableContext = ctx
 
   override def toString: String = toSequent.toSigRelativeString( context )
 }