+++ core/src/main/scala/at/logic/gapt/proofs/resolution/structuralCNF.scala
@@ -1,6 +1,7 @@
 package at.logic.gapt.proofs.resolution
 
 import at.logic.gapt.expr._
+import at.logic.gapt.expr.hol.SkolemFunctions
 import at.logic.gapt.proofs._
 import at.logic.gapt.utils.NameGenerator
 
@@ -8,11 +9,10 @@ import scala.collection.mutable
 
 object structuralCNF {
   def apply(
-
     endSequent:        HOLSequent,
     propositional:     Boolean    = false,
     structural:        Boolean    = true,
-    bidirectionalDefs: Boolean    = false ): Set[ResolutionProof] = {
+    bidirectionalDefs: Boolean    = false )( implicit ctx: MutableContext = MutableContext.guess( endSequent ) ): Set[ResolutionProof] = {
     if ( !propositional )
       require( freeVariables( endSequent ).isEmpty, "end-sequent has free variables" )
 
@@ -25,8 +25,8 @@ object structuralCNF {
     proofs:            Iterable[ResolutionProof],
     propositional:     Boolean                   = false,
     structural:        Boolean                   = true,
-    bidirectionalDefs: Boolean                   = false ): Set[ResolutionProof] = {
-    val clausifier = new Clausifier( propositional, structural, bidirectionalDefs, rename.awayFrom( proofs.flatMap( containedNames( _ ) ) ) )
+    bidirectionalDefs: Boolean                   = false )( implicit ctx: MutableContext = MutableContext.guess( proofs ) ): Set[ResolutionProof] = {
+    val clausifier = new Clausifier( propositional, structural, bidirectionalDefs, ctx, ctx.newNameGenerator )
     proofs foreach clausifier.expand
     clausifier.cnf.toSet
   }
@@ -35,10 +35,16 @@ object structuralCNF {
 class Clausifier(
     propositional: Boolean, structural: Boolean,
     bidirectionalDefs: Boolean,
+    ctx:               MutableContext,
     nameGen:           NameGenerator ) {
   val cnf = mutable.Set[ResolutionProof]()
   val defs = mutable.Map[Expr, HOLAtomConst]()
   val skConsts = mutable.Map[Expr, Const]()
+
+  for ( ( skC, skD ) <- ctx.get[SkolemFunctions].skolemDefs )
+    skConsts( skD ) = skC
+  for ( ( df: HOLAtomConst, by ) <- ctx.definitions )
+    defs( by ) = df
 
   def mkSkolemSym() = nameGen.freshWithIndex( "s" )
   def mkAbbrevSym() = nameGen.freshWithIndex( "D" )
@@ -47,8 +53,11 @@ class Clausifier(
     val fvs = freeVariables( f ).toSeq
     val skolemizedFormula = Abs( fvs, f )
     val skolemConst = skConsts.getOrElseUpdate(
-      skolemizedFormula,
-      Const( mkSkolemSym(), FunctionType( x.ty, fvs map { _.ty } ) ) )
+      skolemizedFormula, {
+      val c = Const( mkSkolemSym(), FunctionType( x.ty, fvs map { _.ty } ) )
+      ctx += Context.SkolemFun( c, skolemizedFormula )
+      c
+    } )
     ( skolemConst( fvs: _* ), skolemizedFormula )
   }
 
@@ -138,8 +147,11 @@ class Clausifier(
     val definedFormula = Abs( fvs, f )
     val alreadyDefined = defs isDefinedAt definedFormula
     val const = defs.getOrElseUpdate(
-      definedFormula,
-      HOLAtomConst( mkAbbrevSym(), fvs map { _.ty }: _* ) )
+      definedFormula, {
+      val c = HOLAtomConst( mkAbbrevSym(), fvs map { _.ty }: _* )
+      ctx += Definition( c, definedFormula )
+      c
+    } )
     if ( !alreadyDefined ) {
       val defn = fvs.foldLeft[ResolutionProof]( Defn( const, definedFormula ) )( AllR( _, Suc( 0 ), _ ) )
       if ( i.isAnt || bidirectionalDefs ) expand( AndR2( defn, Suc( 0 ) ) )