+++ core/src/main/scala/at/logic/gapt/provers/eprover/eprover.scala
@@ -6,14 +6,14 @@ import at.logic.gapt.expr._
 import at.logic.gapt.formats.StringInputFile
 import at.logic.gapt.formats.tptp.{ TPTPFOLExporter, TptpProofParser, tptpToString }
 import at.logic.gapt.proofs.resolution.ResolutionProof
-import at.logic.gapt.proofs.{ FOLClause, HOLClause }
+import at.logic.gapt.proofs.{ FOLClause, HOLClause, MutableContext }
 import at.logic.gapt.proofs.sketch.RefutationSketchToResolution
 import at.logic.gapt.provers.{ ResolutionProver, renameConstantsToFi }
-import at.logic.gapt.utils.{ ExternalProgram, runProcess, withTempFile }
+import at.logic.gapt.utils.{ ExternalProgram, Maybe, runProcess, withTempFile }
 
 object EProver extends EProver
 class EProver extends ResolutionProver with ExternalProgram {
-  override def getResolutionProof( seq: Traversable[HOLClause] ): Option[ResolutionProof] =
+  override def getResolutionProof( seq: Traversable[HOLClause] )( implicit ctx: Maybe[MutableContext] ): Option[ResolutionProof] =
     renameConstantsToFi.wrap( seq.toSeq )(
       ( renaming, cnf: Seq[HOLClause] ) => {
         val labelledCNF = cnf.zipWithIndex.map { case ( clause, index ) => s"formula$index" -> clause.asInstanceOf[FOLClause] }.toMap