+++ tests/src/test/scala/at/logic/gapt/proofs/lk/InductionEliminationTests.scala
@@ -4,7 +4,7 @@ import at.logic.gapt.expr._
 import at.logic.gapt.examples.tip.isaplanner.{ prop_08, prop_15, prop_59 }
 import at.logic.gapt.expr.Substitution
 import at.logic.gapt.formats.tip.TipSmtParser
-import at.logic.gapt.proofs.{ Context, Sequent, SequentMatchers }
+import at.logic.gapt.proofs.{ Context, MutableContext, Sequent, SequentMatchers }
 import at.logic.gapt.proofs.gaptic.{ Lemma, ProofState, allR, cut, escargot, induction, insert, refl, rewrite }
 import org.specs2.mutable.Specification
 
@@ -60,7 +60,7 @@ class InductionEliminationTests extends Specification with SequentMatchers {
   }
 
   "all inductions should be eliminated" in {
-    implicit var ctx = Context()
+    implicit var ctx: MutableContext = MutableContext.default()
     ctx += Context.InductiveType( "nat", hoc"0: nat", hoc"s:nat>nat" )
     ctx += hoc"'+': nat>nat>nat"