+++ RELEASE-NOTES.md
@@ -3,6 +3,7 @@
 ## Version 2.8 (unreleased)
 
 * Support for EProver 2.0
+* MutableContext now keeps track of automatically generated Skolem functions
 
 ## Version 2.7 (released on 2017-07-05)